chore: moving sync logic out of sp client
